update privacy-1 demo
